Small fixes to import. Improved README
